LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCPeerConnection-mandatory-getStats.https-expected.txt


Harness Error (TIMEOUT), message = null

FAIL getStats succeeds assert_equals: Test error. String. expected "string" but got "undefined"
TIMEOUT RTCRtpStreamStats's ssrc Test timed out
NOTRUN RTCRtpStreamStats's kind 
NOTRUN RTCRtpStreamStats's transportId 
NOTRUN RTCRtpStreamStats's codecId 
NOTRUN RTCReceivedRtpStreamStats's packetsReceived 
NOTRUN RTCReceivedRtpStreamStats's packetsLost 
NOTRUN RTCReceivedRtpStreamStats's jitter 
NOTRUN RTCReceivedRtpStreamStats's packetsDiscarded 
NOTRUN RTCInboundRtpStreamStats's trackId 
NOTRUN RTCInboundRtpStreamStats's receiverId 
NOTRUN RTCInboundRtpStreamStats's remoteId 
NOTRUN RTCInboundRtpStreamStats's framesDecoded 
NOTRUN RTCRemoteInboundRtpStreamStats's localId 
NOTRUN RTCRemoteInboundRtpStreamStats's roundTripTime 
NOTRUN RTCSentRtpStreamStats's packetsSent 
NOTRUN RTCSentRtpStreamStats's bytesSent 
NOTRUN RTCOutboundRtpStreamStats's trackId 
NOTRUN RTCOutboundRtpStreamStats's senderId 
NOTRUN RTCOutboundRtpStreamStats's remoteId 
NOTRUN RTCOutboundRtpStreamStats's framesEncoded 
NOTRUN RTCRemoteOutboundRtpStreamStats's localId 
NOTRUN RTCRemoteOutboundRtpStreamStats's remoteTimestamp 
NOTRUN RTCPeerConnectionStats's dataChannelsOpened 
NOTRUN RTCPeerConnectionStats's dataChannelsClosed 
NOTRUN RTCDataChannelStats's protocol 
NOTRUN RTCDataChannelStats's dataChannelIdentifier 
NOTRUN RTCDataChannelStats's state 
NOTRUN RTCDataChannelStats's messagesSent 
NOTRUN RTCDataChannelStats's bytesSent 
NOTRUN RTCDataChannelStats's messagesReceived 
NOTRUN RTCDataChannelStats's bytesReceived 
NOTRUN RTCMediaStreamStats's streamIdentifer 
NOTRUN RTCMediaStreamStats's trackIds 
NOTRUN RTCMediaHandlerStats's trackIdentifier 
NOTRUN RTCMediaHandlerStats's remoteSource 
NOTRUN RTCMediaHandlerStats's ended 
NOTRUN RTCAudioHandlerStats's audioLevel 
NOTRUN RTCVideoHandlerStats's frameWidth 
NOTRUN RTCVideoHandlerStats's frameHeight 
NOTRUN RTCVideoHandlerStats's framesPerSecond 
NOTRUN RTCVideoSenderStats's framesSent 
NOTRUN RTCVideoReceiverStats's framesReceived 
NOTRUN RTCVideoReceiverStats's framesDecoded 
NOTRUN RTCVideoReceiverStats's framesDropped 
NOTRUN RTCVideoReceiverStats's partialFramesLost 
NOTRUN RTCCodecStats's payloadType 
NOTRUN RTCCodecStats's codecType 
NOTRUN RTCCodecStats's clockRate 
NOTRUN RTCCodecStats's channels 
NOTRUN RTCCodecStats's sdpFmtpLine 
NOTRUN RTCTransportStats's bytesSent 
NOTRUN RTCTransportStats's bytesReceived 
NOTRUN RTCTransportStats's rtcpTransportStatsId 
NOTRUN RTCTransportStats's selectedCandidatePairId 
NOTRUN RTCTransportStats's localCertificateId 
NOTRUN RTCTransportStats's remoteCertificateId 
NOTRUN RTCIceCandidatePairStats's transportId 
NOTRUN RTCIceCandidatePairStats's localCandidateId 
NOTRUN RTCIceCandidatePairStats's remoteCandidateId 
NOTRUN RTCIceCandidatePairStats's state 
NOTRUN RTCIceCandidatePairStats's priority 
NOTRUN RTCIceCandidatePairStats's nominated 
NOTRUN RTCIceCandidatePairStats's bytesSent 
NOTRUN RTCIceCandidatePairStats's bytesReceived 
NOTRUN RTCIceCandidatePairStats's totalRoundTripTime 
NOTRUN RTCIceCandidatePairStats's currentRoundTripTime 
NOTRUN RTCIceCandidateStats's address 
NOTRUN RTCIceCandidateStats's port 
NOTRUN RTCIceCandidateStats's protocol 
NOTRUN RTCIceCandidateStats's candidateType 
NOTRUN RTCIceCandidateStats's url 
NOTRUN RTCCertificateStats's fingerprint 
NOTRUN RTCCertificateStats's fingerprintAlgorithm 
NOTRUN RTCCertificateStats's base64Certificate 
NOTRUN RTCCertificateStats's issuerCertificateId 
